Change element creation
Creates element matrix as required, instead of creating and storing
files for each of them.

Material properties are not fixed anymore, and follow what is present on
the TPD file.

Created a separate file for Q4a5B.

Removed files which will not be necessary anymore (matlcons.py and
recreate_all.py).