Fully implement 2D heat support
Heat will not use expand(), as it would not make physical sense.
Messing with the global coefficient would be interesting, but it depends
mostly on heat transfer area, notably from heat input and heat output.

Applying the same calculations as for physical stress would lead to a
topology-specific material thermal conductivity, which tends to be
smaller than the one used.

But nonetheless, the topology optimization algorithm tends to offset the
material addition made by expand() anyway, so there is not much point to
it.

Using pathfinding however is still helpful to increase calculation speed
for each iteration, and even more without the use of expand(), so the
preprocessing algorithm is still useful in some way for heat conduction
optimization.